Refactor method for running format 2 workflows into populators.py for reuse.
